Module bevy_trackball::prelude

Prelude to get started quickly.

Re-exports

pub use super::TrackballCamera;

pub use super::TrackballController;

pub use super::TrackballEvent;

pub use super::TrackballInput;

pub use super::TrackballPlugin;

pub use super::TrackballSetup;

pub use super::TrackballSystemSet;

pub use super::TrackballVelocity;

pub use super::TrackballViewport;

pub use super::TrackballWheelUnit;

Macros

abs_diff_eq

Approximate equality of using the absolute difference.

abs_diff_ne

Approximate inequality of using the absolute difference.

assert_abs_diff_eq

An assertion that delegates to abs_diff_eq!, and panics with a helpful error on failure.

assert_abs_diff_ne

An assertion that delegates to abs_diff_ne!, and panics with a helpful error on failure.

assert_relative_eq

An assertion that delegates to relative_eq!, and panics with a helpful error on failure.

assert_relative_ne

An assertion that delegates to relative_ne!, and panics with a helpful error on failure.

assert_ulps_eq

An assertion that delegates to ulps_eq!, and panics with a helpful error on failure.

assert_ulps_ne

An assertion that delegates to ulps_ne!, and panics with a helpful error on failure.

relative_eq

Approximate equality using both the absolute difference and relative based comparisons.

relative_ne

Approximate inequality using both the absolute difference and relative based comparisons.

ulps_eq

Approximate equality using both the absolute difference and ULPs (Units in Last Place).

ulps_ne

Approximate inequality using both the absolute difference and ULPs (Units in Last Place).

Structs

Bound

Orthogonal boundary conditions implementing Clamp.

Frame

Frame wrt camera eye and target.

Plane

Plane encoding position with singed bias along unit normal.

Scope

Scope defining enclosing viewing frustum.

Unit

A wrapper that ensures the underlying algebraic entity has a unit norm.

Enums

Delta

Delta transform from initial to final Frame.

Fixed

Fixed quantity wrt field of view.

Traits

AbsDiffEq

Equality that is defined using the absolute difference of two numbers.

Clamp

Clamp wrt abstract boundary conditions of Frame and Scope.

RelativeEq

Equality comparisons between two numbers using both the absolute difference and relative based comparisons.

UlpsEq

Equality comparisons between two numbers using both the absolute difference and ULPs (Units in Last Place) based comparisons.

Type Aliases

Isometry3

A 3-dimensional direct isometry using a unit quaternion for its rotational part.

Point3

A statically sized 3-dimensional column point.

UnitQuaternion

A unit quaternions. May be used to represent a rotation.

Vector3

A stack-allocated, 3-dimensional column vector.