Struct bevy_trackball::prelude::Plane

#[repr(C)]
pub struct Plane<N>
where
    N: Copy + RealField,
{
    pub normal: Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>>,
    pub bias: N,
}

Plane encoding position with singed bias along unit normal.

Realizes plane equation a*x+b*y+c*z+d=0 with unit normal [x, y, z] and signed bias d.

Fields
normal: Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>>

Plane unit normal.

bias: N

Signed bias along unit normal.

Implementations

impl<N> Plane<N>
where N: Copy + RealField,

pub fn new( normal: Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>>, distance: N ) -> Plane<N>

Plane from unit normal and signed distance from the origin.

use trackball::{
    nalgebra::{Point3, Vector3},
    Plane,
};

// Plane intersecting y-axis at `5.0`.
let plane = Plane::new(Vector3::y_axis(), 5.0);

// Origin projected onto plane where plane intersects y-axis.
let point = plane.project_point(&Point3::origin());
assert_eq!(point, Point3::new(0.0, 5.0, 0.0));

// Bias is negated distance.
assert_eq!(plane.distance(), 5.0);
assert_eq!(plane.bias, -5.0);

// Plane intersecting y-axis at `-5.0`.
let plane = Plane::new(Vector3::y_axis(), -5.0);

// Origin projected onto plane where plane intersects y-axis.
let point = plane.project_point(&Point3::origin());
assert_eq!(point, Point3::new(0.0, -5.0, 0.0));

// Bias is negated distance.
assert_eq!(plane.distance(), -5.0);
assert_eq!(plane.bias, 5.0);

pub fn with_point( normal: Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>>, point: &OPoint<N, Const<3>> ) -> Plane<N>

Plane from unit normal with point in plane.

pub fn distance(&self) -> N

Signed orthogonal distance from the origin.

pub fn distance_from(&self, point: &OPoint<N, Const<3>>) -> N

Signed orthogonal distance from point.

pub fn project_point(&self, point: &OPoint<N, Const<3>>) -> OPoint<N, Const<3>>

Projects point onto plane.

pub fn project_axis( &self, axis: &Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>> ) -> Unit<Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>>

Projects axis onto plane.

pub fn project_vector( &self, vector: &Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>> ) -> Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>

Projects vector onto plane.

pub fn angle_between( &self, a: &Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>>, b: &Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>> ) -> N

Singed angle from a to b where both vectors are in the plane.

pub fn rotate_by(self, rot: &Unit<Quaternion<N>>) -> Plane<N>

Rotates plane.

pub fn translate_by( self, vec: &Matrix<N, Const<3>, Const<1>, ArrayStorage<N, 3, 1>> ) -> Plane<N>

Translates plane.

pub fn transform_by(self, iso: &Isometry<N, Unit<Quaternion<N>>, 3>) -> Plane<N>

Transforms plane by direct isometry, i.e., rotation followed by translation.

use core::f64::{
    consts::FRAC_PI_2,
    EPSILON,
};
use trackball::{
    approx::AbsDiffEq,
    nalgebra::{Isometry3, Vector3},
    Plane,
};

// Plane intersecting y-axis at `5.0`.
let plane = Plane::new(Vector3::y_axis(), 5.0)
    .transform_by(&Isometry3::new(
        // Translation by after rotation.
        Vector3::new(-5.0, 0.0, 0.0),
        // Rotation by 90 degrees before translation.
        Vector3::new(0.0, 0.0, FRAC_PI_2),
    ));
// Plane intersecting x-axis at `-10.0`.
assert!(plane.abs_diff_eq(&Plane::new(-Vector3::x_axis(), 10.0), EPSILON));

pub fn cast<M>(self) -> Plane<M>
where M: Copy + RealField, N: SubsetOf<M>,

Casts components to another type, e.g., between f32 and f64.