Type Alias bevy_trackball::prelude::Isometry3

pub type Isometry3<T> = Isometry<T, Unit<Quaternion<T>>, 3>;

A 3-dimensional direct isometry using a unit quaternion for its rotational part.

Because this is an alias, not all its methods are listed here. See the Isometry type too.

Also known as a rigid-body motion, or as an element of SE(3).

Aliased Type

struct Isometry3<T> {
    pub rotation: Unit<Quaternion<T>>,
    pub translation: Translation<T, 3>,
}

Fields
rotation: Unit<Quaternion<T>>

The pure rotational part of this isometry.

translation: Translation<T, 3>

The pure translational part of this isometry.