Struct bevy_trackball::TrackballCamera

pub struct TrackballCamera {
    pub frame: Frame<f32>,
    pub scope: Scope<f32>,
    pub blend: f32,
    pub reset: Frame<f32>,
    pub clamp: Option<Box<dyn Clamp<f32>>>,
    pub group: HashMap<Entity, bool>,
    /* private fields */
}

Trackball camera component mainly defined by Frame and Scope.

Fields
frame: Frame<f32>

Camera frame defining Transform.

Comprises following properties:

target position as trackball center

camera eye rotation on trackball surface (incl. roll, gimbal lock-free using quaternion)

trackball radius

scope: Scope<f32>

Camera scope defining Projection.

Comprises following properties:

field of view angle (default is 45 degrees) and its mode of either Fixed::Ver (default), Fixed::Hor, or Fixed::Upp.

projection mode of either perspective (default) or orthographic (scale preserving)

clip planes either measured from eye (default) or target (object inspection mode)

blend: f32

Blend ratio between 0 (slow) and 1 (fast) for exponential easy-out. Default is 0.25.

Blend ratio aka sharpness or Lerp/Slerp parameter t is exponentially corrected for deviations from reference frame rate of 60 fps.

reset: Frame<f32>

Camera frame to reset to when TrackballInput::reset_key is pressed.

clamp: Option<Box<dyn Clamp<f32>>>

User boundary conditions clamping camera Frame.

Allows to limit target/eye position or minimal/maximal target/eye distance or up rotation.

group: HashMap<Entity, bool>

Additional TrackballController entities to which this camera is sensitive.

It is always sensitive to its own controller if it has one. A mapped value of true will clamp the active controller as well and hence all other cameras of this group whenever this camera is clamped. If false, only this camera is clamped whereas other cameras of this group continue to follow the active controller.

Implementations

impl TrackballCamera

pub fn look_at(target: Vec3, eye: Vec3, up: Vec3) -> Self

Defines camera with target position and eye position inclusive its roll attitude (up).

pub const fn with_scope(self, scope: Scope<f32>) -> Self

Defines scope, see Self::scope.

pub const fn with_blend(self, blend: f32) -> Self

Defines blend ratio, see Self::blend.

pub const fn with_reset(self, reset: Frame<f32>) -> Self

Defines reset frame, see Self::reset.

pub fn with_clamp(self, clamp: impl Clamp<f32>) -> Self

Defines user boundary conditions, see Self::clamp.

pub fn add_controller(self, id: Entity, rigid: bool) -> Self

Adds additional controller to which this camera is sensitive, see Self::group.