Struct bevy_trackball::TrackballInput

pub struct TrackballInput {

    pub velocity: TrackballVelocity,
    pub wheel_unit: TrackballWheelUnit,
    pub focus: bool,
    pub gamer_key: Option<KeyCode>,
    pub ortho_key: Option<KeyCode>,
    pub reset_key: Option<KeyCode>,
    pub first_button: Option<MouseButton>,
    pub first_key: Option<KeyCode>,
    pub first_left_key: Option<KeyCode>,
    pub first_right_key: Option<KeyCode>,
    pub first_up_key: Option<KeyCode>,
    pub first_down_key: Option<KeyCode>,
    pub orbit_button: Option<MouseButton>,
    pub screw_left_key: Option<KeyCode>,
    pub screw_right_key: Option<KeyCode>,
    pub orbit_left_key: Option<KeyCode>,
    pub orbit_right_key: Option<KeyCode>,
    pub orbit_up_key: Option<KeyCode>,
    pub orbit_down_key: Option<KeyCode>,
    pub slide_button: Option<MouseButton>,
    pub slide_left_key: Option<KeyCode>,
    pub slide_right_key: Option<KeyCode>,
    pub slide_up_key: Option<KeyCode>,
    pub slide_down_key: Option<KeyCode>,
    pub slide_far_key: Option<KeyCode>,
    pub slide_near_key: Option<KeyCode>,
    pub scale_in_key: Option<KeyCode>,
    pub scale_out_key: Option<KeyCode>,

}

Trackball controller input mappings and settings.

Fields
velocity: TrackballVelocity

Trackball velocity for time-based input like pressed keys.

wheel_unit: TrackballWheelUnit

Wheel unit for coherent scaling. Default is 24 clicks per turn.

Device dependent setting as mouse wheel events usually lack a reference unit.

focus: bool

Enables focus operation. Default is true.

Whether to slide towards mouse or single-finger touch position when Self::orbit_button is just pressed and released again or single-finger gesture is just started and ended again. Moving the cursor/finger slightly between pressed/started and released/ended events discards the focus operation in favor of the orbit operation.

gamer_key: Option<KeyCode>

Key used to toggle esdf/wasd mapping. Default is KeyCode::KeyM.

ortho_key: Option<KeyCode>

Key used to toggle projection mode. Default is KeyCode::KeyP.

reset_key: Option<KeyCode>

Key used to reset frame. Default is KeyCode::Enter.

first_button: Option<MouseButton>

Mouse button used to look around. Default is MouseButton::Middle.

first_key: Option<KeyCode>

Key used to look around with single-finger touch. Default is KeyCode::ShiftLeft.

first_left_key: Option<KeyCode>

Key used to look left. Default is KeyCode::ArrowLeft.

first_right_key: Option<KeyCode>

Key used to look right. Default is KeyCode::ArrowRight.

first_up_key: Option<KeyCode>

Key used to look up. Default is KeyCode::ArrowUp.

first_down_key: Option<KeyCode>

Key used to look down. Default is KeyCode::ArrowDown.

orbit_button: Option<MouseButton>

Mouse button used to orbit camera. Default is MouseButton::Left.

screw_left_key: Option<KeyCode>

Key used to screw/roll left. Default is KeyCode::KeyU.

screw_right_key: Option<KeyCode>

Key used to screw/roll right. Default is KeyCode::KeyO.

orbit_left_key: Option<KeyCode>

Key used to orbit left. Default is KeyCode::KeyJ.

orbit_right_key: Option<KeyCode>

Key used to orbit right. Default is KeyCode::KeyL.

orbit_up_key: Option<KeyCode>

Key used to orbit up. Default is KeyCode::KeyI.

orbit_down_key: Option<KeyCode>

Key used to orbit down. Default is KeyCode::KeyK.

slide_button: Option<MouseButton>

Mouse button used to slide camera. Default is MouseButton::Right.

slide_left_key: Option<KeyCode>

Key used to slide left. Default is KeyCode::KeyS.

slide_right_key: Option<KeyCode>

Key used to slide right. Default is KeyCode::KeyF.

slide_up_key: Option<KeyCode>

Key used to slide up. Default is KeyCode::KeyE.

slide_down_key: Option<KeyCode>

Key used to slide up. Default is KeyCode::KeyD.

slide_far_key: Option<KeyCode>

Key used to slide far. Default is KeyCode::KeyG.

slide_near_key: Option<KeyCode>

Key used to slide near. Default is KeyCode::KeyV.

scale_in_key: Option<KeyCode>

Key used to scale/zoom in. Default is KeyCode::KeyH.

scale_out_key: Option<KeyCode>

Key used to scale/zoom out. Default is KeyCode::KeyN.

Implementations

impl TrackballInput

pub fn map_esdf(&mut self)

Maps esdf/gv to slide operations.

Key	Operation

e	Slides up.

s	Slides left.

d	Slides down.

f	Slides right.

g	Slides far (in/forward).

v	Slides near (out/backward).

This mapping is symmetric to the ijkl/hn orbit mapping but less intuitive to gamers compared with Self::map_wasd.

pub fn map_wasd(&mut self)

Maps wasd/Space/ControlLeft to slide operations.

Key	Operation

w	Slides far (in/forward).

a	Slides left.

s	Slides near (out/backward).

d	Slides right.

Space	Slides up (jump).

ControlLeft	Slides down (crouch).

This mapping isn’t symmetric to the ijkl/hn orbit mapping but more intuitive to gamers compared with Self::map_esdf.